move metamask wallet creation logic to the metamask service
