+++ src/components/Wallet/MetamaskWallet/index.tsx
@@ -1,23 +1,17 @@
-import { Signer } from '@polkadot/api/types';
-import { SignerPayloadJSON, SignerPayloadRaw, SignerResult } from '@polkadot/types/types';
-import { WalletAccount } from '@talismn/connect-wallets';
 import { useCallback, useEffect, useState } from 'preact/compat';
 import { Modal } from 'react-daisyui';
 import { GlobalState, useGlobalState } from '../../../GlobalStateProvider';
 import logo from '../../../assets/metamask-wallet.png';
-import { initiatePolkadotSnap, installPolkadotSnap, tenantToSnapNetwork } from '../../../services/metamask/metamask';
+import {
+  ExtensionAccount,
+  buildWalletAccount,
+  initiateMetamaskInjectedAccount,
+} from '../../../services/metamask/metamask';
 import { PublicKey } from '../../PublicKey';
 
 export type MetamaskWalletProps = {
   setWalletAccount: GlobalState['setWalletAccount'];
 };
-
-interface ExtensionAccount {
-  address: string;
-  name: string;
-  source: string;
-  signer: Signer;
-}
 
 const MetamaskWallet = ({ setWalletAccount }: MetamaskWalletProps) => {
   const [openModal, setOpenModal] = useState<boolean>(false);
@@ -26,39 +20,8 @@ const MetamaskWallet = ({ setWalletAccount }: MetamaskWalletProps) => {
   const { tenantName } = useGlobalState();
 
   const onClick = useCallback(async () => {
-    const provider = await initiatePolkadotSnap(tenantToSnapNetwork(tenantName));
-    if (!provider.isSnapInstalled) {
-      const installResult = await installPolkadotSnap(tenantToSnapNetwork(tenantName));
-      if (!installResult) {
-        console.error('Something went wrong, snap could not be installed.');
-        setOpenModal(false);
-        return;
-      }
-    }
-    if (provider.snap) {
-      const api = provider.snap.getMetamaskSnapApi();
-      const injectedMetamaskAccount: ExtensionAccount = {
-        address: await api.getAddress(),
-        name: 'Metamask Snap',
-        source: 'metamask',
-        signer: {
-          signPayload: async (payload: SignerPayloadJSON) => {
-            const stringResult = await api.signPayloadJSON(payload);
-            // Metamask snap doesn't provide a request Id, but just the hex string, so
-            // adding id: 1 to be compliant with SignerResult
-            return { id: 1, signature: stringResult } as SignerResult;
-          },
-          signRaw: async (raw: SignerPayloadRaw) => {
-            const stringResult = await api.signPayloadRaw(raw);
-            // Metamask snap doesn't provide a request Id, but just the hex string, so
-            // adding id: 1 to be compliant with SignerResult
-            return { id: 1, signature: stringResult } as SignerResult;
-          },
-          update: (id, status) => {
-            console.log('Status update for Id %d: %s', id, status.toHuman());
-          },
-        },
-      };
+    const injectedMetamaskAccount = (await initiateMetamaskInjectedAccount(tenantName)) as ExtensionAccount;
+    if (injectedMetamaskAccount) {
       setAccounts([injectedMetamaskAccount]);
       setOpenModal(true);
     } else {
@@ -68,34 +31,6 @@ const MetamaskWallet = ({ setWalletAccount }: MetamaskWalletProps) => {
   }, [setOpenModal]);
 
   useEffect(() => {
-    async function buildWalletAccount(extAcc: ExtensionAccount) {
-      return {
-        address: extAcc.address,
-        source: extAcc.source,
-        name: extAcc.name,
-        signer: extAcc.signer as WalletAccount['signer'],
-        wallet: {
-          enable: () => undefined,
-          extensionName: 'polkadot-js',
-          title: 'Metamask Wallet',
-          installUrl: 'https://metamask.io/',
-          logo: {
-            src: logo,
-            alt: 'Metamask Wallet',
-          },
-          installed: true,
-          extension: undefined,
-          signer: extAcc.signer,
-          /**
-           * The following methods are tagged as 'Unused' since they are only required by the @talisman package,
-           * which we are not using to handle this wallet connection.
-           */
-          getAccounts: () => Promise.resolve([]), // Unused
-          subscribeAccounts: () => undefined, // Unused
-          transformError: (err: Error) => err, // Unused
-        },
-      };
-    }
     if (selectedAccount) {
       buildWalletAccount(selectedAccount)
         .then((account) => setWalletAccount(account))
